+++ src/python/zensols/showfile/app.py
@@ -65,8 +65,6 @@ class Application(object):
             for loc in pres.locators:
                 loc.coerce_type(locator_type)
         try:
-            for loc in pres.locators:
-                loc.validate()
             self.browser_manager.show(pres)
         except ShowFileError as e:
             raise ApplicationError(str(e)) from e